Fix the default num files and max file size
default num files was not set and if run time set it to 0
(or below 20), it would run into a deadlock (read waiting for write,
write waiting for read).

default max file size was set to -1 instead of the correct number
of bits (30).

help text fix for the max num files and several other small fixes.