Add a script to simplify the run
ql-fstest is supposed to be run with multipe processes
per host. It would be possible to add more parallism into
ql-fstest itself, but then there is not really a good
reason for that, if one just needs to start it multiple
times. I was using a similar set of scripts for ages
in another test project and this just adds it to this
project internally.

example to run for 60s with 2 processes

./run-ql-fstest.sh -t60 -e -n2 -d /tmp2/test


once that completes, or in between, one can check for errors
using

./check-ql-fstest.sh /tmp2/test